backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
httpx>=0.28.0
pandas>=2.2.0
numpy>=2.0.0
scikit-learn>=1.6.0
xgboost>=2.1.0
joblib>=1.4.0
pydantic>=2.10.0
python-dotenv>=1.0.0
supabase>=2.0.0
